Added timer and cleaned up messages
