+++ Cargo.toml
@@ -16,3 +16,4 @@ regex = "1"
 lazy_static = "1"
 glob = "0.3"
 clap = { version = "4.1.8", features = ["derive"] }
+
