+++ src/parse_and_printout/parse.rs
@@ -63,53 +63,18 @@ fn check_reserved(word: String) -> bool {
     false
 }
 
+// extern crate test;
+
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_get_classes_or_ids() {
-        let class_or_id =
-            ".testClass {}\n#testId {}\n.errorClass {\n@value test;".get_classes_or_ids();
-        let class_or_id_expected = [".testClass", "#testId", "@value test"];
-        assert_eq!(class_or_id, class_or_id_expected)
-    }
-
-    #[test]
-    fn split_string() {
-        let (first_char, remainder) = "test".split_first_char();
-        assert_eq!(first_char, "t");
-        assert_eq!(remainder, "est")
-    }
-
-    #[test]
-    fn test_remove_comments() {
-        let text = "/*\n.commentClass {}\n*/\n.test{}".to_string();
-        let parsed_text = &text.remove_comments();
-        assert_eq!(parsed_text, &"\n.test{}")
-    }
-
-    #[test]
-    fn test_remove_modifiers() {
-        let parsed_id = "#test:modifiers".remove_modifiers();
-        let parsed_val = "@value test".remove_modifiers();
-        assert_eq!(
-            (parsed_id, parsed_val),
-            ("test".to_string(), "test".to_string())
-        )
-    }
+    // use test::Bencher;
 
     #[test]
     fn check_reserved_names() {
         let res_true = check_reserved("any".to_string());
         let res_false = check_reserved("test".to_string());
         assert_eq!((res_true, res_false), (true, false))
-    }
-
-    #[test]
-    fn camel_case() {
-        let name = "Hello-world".camel_case_converter();
-        assert_eq!(name, "helloWorld");
     }
 
     #[test]
@@ -137,6 +102,19 @@ mod tests {
         }
         assert_eq!((diff, file_data.1), (false, file_data_expected.1))
     }
+
+    // #[bench]
+    // fn bench_parse_file_data(b: &mut Bencher) {
+    //     b.iter(|| {
+    //         parse_file_data(
+    //             &"./test/test.module.css".to_string(),
+    //             &ModFlags {
+    //                 camel_case_flag: false,
+    //                 out_dir: &String::new(),
+    //             },
+    //         )
+    //     })
+    // }
 
     #[test]
     fn test_get_file_recursive_data() {
